+++ lib/benchee/config.ex
@@ -27,12 +27,14 @@ defmodule Benchee.Config do
     * print      - a map from atoms to `true` or `false` to configure if the
     output identified by the atom will be printed. All options are enabled by
     default (true). Options are:
-      * `:comparison` - the comparison of the different benchmarking jobs (x
+      * `:benchmarking`  - print when Benchee starts benchmarking a new job
+      (Benchmarking name ..)
+      * `:comparison`    - the comparison of the different benchmarking jobs (x
       times slower than)
       * `:configuration` - a summary of configured benchmarking options
       including estimated total run time is printed before benchmarking starts
-      * `:fast_warning` - warnings are displayed if functions are executed too
-        fast leading to inaccurate measures
+      * `:fast_warning`   - warnings are displayed if functions are executed
+      too fast leading to inaccurate measures
 
   ## Examples
 
@@ -44,7 +46,12 @@ defmodule Benchee.Config do
             time: 5_000_000,
             warmup: 2_000_000,
             formatters: [&Benchee.Formatters.Console.output/1],
-            print: %{fast_warning: true, comparison: true, configuration: true}
+            print: %{
+              benchmarking: true,
+              fast_warning: true,
+              comparison: true,
+              configuration: true
+            }
           },
         jobs: %{}
       }
@@ -57,7 +64,12 @@ defmodule Benchee.Config do
             time: 1_000_000,
             warmup: 200_000.0,
             formatters: [&Benchee.Formatters.Console.output/1],
-            print: %{fast_warning: true, comparison: true, configuration: true}
+            print: %{
+              benchmarking: true,
+              fast_warning: true,
+              comparison: true,
+              configuration: true
+            }
           },
         jobs: %{}
       }
@@ -70,7 +82,12 @@ defmodule Benchee.Config do
             time: 1_000_000,
             warmup: 200_000.0,
             formatters: [&IO.puts/2],
-            print: %{fast_warning: false, comparison: true, configuration: true}
+            print: %{
+              benchmarking: true,
+              fast_warning: false,
+              comparison: true,
+              configuration: true
+            }
           },
         jobs: %{}
       }
@@ -81,9 +98,10 @@ defmodule Benchee.Config do
     warmup:     2,
     formatters: [&Benchee.Formatters.Console.output/1],
     print:      %{
-                  fast_warning: true,
-                  comparison: true,
-                  configuration: true
+                  benchmarking:  true,
+                  comparison:    true,
+                  configuration: true,
+                  fast_warning:  true
                 }
   }
   @time_keys [:time, :warmup]