+++ samples/deactivate_output.exs
@@ -3,9 +3,10 @@
 Benchee.run(%{
                 time: 1,
                 print: %{
-                  fast_warning: false,
-                  comparison: false,
-                  configuration: false
+                  benchmarking:  false,
+                  comparison:    false,
+                  configuration: false,
+                  fast_warning:  false
                 }
               },
             %{"fast" => fn -> 1 + 1 end,
@@ -13,9 +14,7 @@ Benchee.run(%{
             })
 
 # tobi@airship ~/github/benchee $ mix run samples/deactivate_output.exs
-# Benchmarking also...
-# Benchmarking fast...
 #
 # Name           ips        average    deviation         median
-# fast   68351710.36       0.0146μs    (±23.14%)       0.0140μs
-# also   67527366.91       0.0148μs    (±27.62%)       0.0140μs
+# fast   67196569.41       0.0149μs    (±24.72%)       0.0140μs
+# also   66268711.31       0.0151μs    (±24.23%)       0.0140μs